AppVeyor: update compiler versions, use Msys2, and Windows fixes (#593)
AppVeyor was using Visual Studio 2015 along with old versions of MinGW. Update to the latest AppVeyor provides and also add an MSYS2 build test using MinGW which will use the bleeding edge version.

When researching #590 this also uncovered a bug in cmake not properly detecting if_indextoname() on windows. This has been corrected as well as the underlying issue reported in #590.

Fix By: Brad House (@bradh352) and Jonas Kvinge (@jonaski)

+++ src/lib/ares__timeval.c
@@ -39,8 +39,8 @@ struct timeval ares__tvnow(void)
   */
   struct timeval now;
   DWORD          milliseconds = GetTickCount();
-  now.tv_sec                  = milliseconds / 1000;
-  now.tv_usec                 = (milliseconds % 1000) * 1000;
+  now.tv_sec                  = (long)milliseconds / 1000;
+  now.tv_usec                 = (long)(milliseconds % 1000) * 1000;
   return now;
 }
