+++ src/lib/inet_ntop.c
@@ -178,13 +178,13 @@ static const char *inet_ntop6(const unsigned char *src, char *dst, size_t size)
     if (i == 6 && best.base == 0 &&
         (best.len == 6 || (best.len == 7 && words[7] != 0x0001) ||
          (best.len == 5 && words[5] == 0xffff))) {
-      if (!inet_ntop4(src + 12, tp, sizeof(tmp) - (tp - tmp))) {
+      if (!inet_ntop4(src + 12, tp, sizeof(tmp) - (size_t)(tp - tmp))) {
         return (NULL);
       }
       tp += ares_strlen(tp);
       break;
     }
-    tp += snprintf(tp, sizeof(tmp) - (tp - tmp), "%x", words[i]);
+    tp += snprintf(tp, sizeof(tmp) - (size_t)(tp - tmp), "%x", words[i]);
   }
   /* Was it a trailing run of 0x00's? */
   if (best.base != -1 &&
